rename snapshot_download.py -> _snapshot_download.py
